Изменение опроса кнопки, исправление ошибки, добавление примерных кадров

+++ LCD_gauge/Open429I-C-Demo/Project/STemWin_RTOS _7inch/Core/User/proto.h
@@ -11,7 +11,7 @@ typedef struct
 		uint16_t fuelLevel;
 		uint16_t oilPressure;
 		uint16_t coolantTemperature;
-		uint16_t pictoState;
+		uint32_t pictoState;
 } stProtocolData;
 #pragma pack() 
 void Proto_Init(void);
